backend/requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
Django==4.1.7
django-environ==0.10.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
PyMySQL==1.0.2
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
